+++ packages/solid/src/components/file-upload/tests/basic.tsx
@@ -14,6 +14,7 @@ export const ComponentUnderTest = (props: FileUpload.RootProps) => (
               <FileUpload.Label>Drag your file(s) here</FileUpload.Label>
             </FileUpload.Dropzone>
             <FileUpload.Trigger>Choose file(s)</FileUpload.Trigger>
+            <FileUpload.ClearTrigger>Clear</FileUpload.ClearTrigger>
             <FileUpload.ItemGroup>
               <For each={fileUpload().acceptedFiles}>
                 {(file) => (